2008-09-18  Sancho Lerena <slerena@artica.es>
	* install.php: Fixed a bad description string.

	* pandoradb.sql: Precission for the float in post_process was
	insignificant (2 decimals), not valid to transform from bytes
	to MB for example, altered to have 13 decimals. Thanks to 
	darkgate for this bug report.



git-svn-id: https://svn.code.sf.net/p/pandora/code/trunk@1099 c3f86ba8-e40f-0410-aaad-9ba5e7f4b01f

+++ pandora_console/pandoradb.sql
@@ -153,7 +153,7 @@ CREATE TABLE IF NOT EXISTS `tagente_modulo` (
   `plugin_pass` varchar(250) default '',
   `plugin_parameter` text,
   `id_plugin` int(10) default '0',
-  `post_process` double(18,2) default NULL,
+  `post_process` double(18,13) default NULL,
   `prediction_module` bigint(14) default '0',
   `max_timeout` int(4) unsigned default '0',
   PRIMARY KEY  (`id_agente_modulo`),
